feat: Add models for rest of migrations and restructure to contexts
Use contexts from phx 1.3 and change structure to have contexts. Also
change scraper to user rust for processing

+++ server/apps/models/priv/repo/migrations/20170319055028_create_combat_average_statistics.exs
@@ -3,16 +3,16 @@ defmodule Models.Repo.Migrations.CreateCombatAverageStatistics do
 
   def change do
     create table(:combat_average_statistics) do
-      add :solo_kills_average, :integer
-      add :eliminations_average, :integer
+      add :solo_kills_average, :decimal
+      add :eliminations_average, :decimal
       add :time_spent_on_file_average, :integer
-      add :deaths_average, :integer
+      add :deaths_average, :decimal
       add :damage_done_average, :integer
       add :objective_time_average, :integer
-      add :objective_kills_average, :integer
-      add :healing_done_average, :integer
-      add :final_blows_average, :integer
-      add :melee_final_blows_average, :integer
+      add :objective_kills_average, :decimal
+      add :healing_done_average, :decimal
+      add :final_blows_average, :decimal
+      add :melee_final_blows_average, :decimal
     end
   end
 end
