+++ server/apps/models/priv/repo/migrations/20170319055036_create_combat_lifetime_statistics.exs
@@ -5,7 +5,7 @@ defmodule Models.Repo.Migrations.CreateCombatLifetimeStatistics do
     create table(:combat_lifetime_statistics) do
       add :solo_kills, :integer
       add :eliminations, :integer
-      add :eliminations_per_life, :integer
+      add :eliminations_per_life, :decimal
       add :time_spent_on_fire, :integer
       add :deaths, :integer
       add :objective_kills, :integer
